+++ src/app/admin/admin.component.html
@@ -1,0 +1,13 @@
+
+<div>
+  <label>
+    description
+    <input type="text" [(ngModel)]="description">
+  </label>
+  <button type="button" (click)="sendInviteKakao();">초대발송</button>
+</div>
+
+<div>
+  <button type="button" (click)="removeGuest();">guest 삭제</button>
+  <button type="button" (click)="removeBdData();">bdData 삭제</button>
+</div>
