+++ src/app/admin/admin.component.scss
@@ -1,0 +1,10 @@
+div {
+  width: 100%;
+  margin: 30px 0;
+  &:first-child{
+    margin-top: 100px;
+  }
+  &:last-child{
+    margin-bottom: 200px;
+  }
+}
