+++ src/app/select/select.component.html
@@ -9,5 +9,3 @@
     <button type="button" (click)="setGuest();">😗 시 작 😗</button>
   </div>
 </div>
-
-<button class="removeBtn" (click)="removeGuest();">re</button>
