+++ src/app/select/select.component.scss
@@ -26,11 +26,3 @@
     padding: 3px 20px;
   }
 }
-
-.removeBtn {
-  position: absolute;
-  bottom: 0;
-  right: 0;
-  border: none;
-  background: none;
-}
