+++ src/app/select/select.component.ts
@@ -38,8 +38,5 @@ export class SelectComponent implements OnInit {
     }
   }
 
-  public removeGuest() {
-    this.cookieService.delete('bdGuest');
-  }
 
 }
